moved tool file into tools in core
