+++ core/tools/object_detection_tool.py
